# backend/openapi.yaml
openapi: 3.0.1
info:
  title: AguaVigia CTG — API publica
  description: |-
    Monitoreo ciudadano del acueducto de Cartagena de Indias.

    Plataforma ciudadana e independiente: no esta afiliada a Aguas de Cartagena
    S.A. E.S.P. ni a ninguna entidad distrital.
  contact:
    name: Equipo AguaVigia CTG — Tecnologico Comfenalco
  license:
    name: Proyecto de aula
  version: 1.0.0
servers:
- url: /
  description: Mismo origen que sirve la API
...
paths:
...
  /api/veedor/segundo-factor/confirmacion:
    post:
      tags:
      - Veedor - Segundo factor
      summary: Confirmar el alta con un codigo de la app
      description: |-
        Devuelve una sesion nueva de alcance COMPLETO. Es lo que permite que un ADMIN
        recien sembrado pase de su sesion restringida al panel sin volver a escribir la
        clave que acaba de escribir.
      operationId: confirmar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCodigo"
        required: true
      responses:
        "200":
          description: Segundo factor activo; sesion nueva emitida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "401":
          description: El codigo no coincide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "409":
          description: No hay un alta en curso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
...
  /api/veedor/cortes:
    get:
      tags:
      - Veedor - Cortes
      summary: Listar los cortes que afectan a un sector
      operationId: listarPorSector
      parameters:
      - name: sectorId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorteRespuesta"
    post:
      tags:
      - Veedor - Cortes
      summary: Registrar un corte oficial
      description: "Sectores afectados, inicio, fin prometido y causa (RF016). Origen\
        \ VEEDOR."
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCorte"
        required: true
      responses:
        "201":
          description: Corte registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "400":
          description: Datos inválidos o algún sector no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/reportes:
    post:
      tags:
      - Reportes
      summary: Registrar un reporte ciudadano
      description: |-
        Sin registro ni cuenta (RF005). Limita automáticamente los reportes por
        dispositivo en la ventana vigente (RF006) — ver 429. La coordenada es opcional,
        solo si el usuario autorizó compartir su ubicación (RF007).
      operationId: registrar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudReporte"
        required: true
      responses:
        "201":
          description: Reporte registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "400":
          description: Sector inexistente o tipo de reporte inválido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: El dispositivo superó el límite de reportes para este sector
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/reportes/{id}/confirmar:
    post:
      tags:
      - Reportes
      summary: Confirmar un reporte
      description: Permite a otro vecino confirmar un reporte ciudadano (M11).
      operationId: confirmar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudConfirmar"
        required: true
      responses:
        "200":
          description: Reporte confirmado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "400":
          description: Error en la solicitud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
...
  /api/cuentas/verificacion:
    post:
      tags:
      - Cuentas
      summary: Confirmar el correo con el token del enlace
      description: Pasa la cuenta a PENDIENTE_APROBACION. Sigue sin poder entrar hasta
        que un ADMIN la apruebe.
      operationId: verificarCorreo
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Correo confirmado
        "400":
          description: "Enlace invalido, vencido o ya usado"
  /api/cuentas/restablecimiento:
    post:
      tags:
      - Cuentas
      summary: Pedir el enlace para restablecer la clave
      description: "Responde 202 siempre, exista o no la cuenta. Ver el javadoc de\
        \ esta clase."
      operationId: pedirRestablecimiento
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudRestablecer"
        required: true
      responses:
        "202":
          description: Accepted
  /api/cuentas/registro:
    post:
      tags:
      - Cuentas
      summary: Solicitar una cuenta del panel
      description: |-
        Crea la cuenta en PENDIENTE_VERIFICACION y envia el enlace de confirmacion.
        Registrarse no concede ningun permiso: hace falta verificar el correo y que un
        ADMIN apruebe. Responde 202 aunque el correo ya tenga cuenta, para no revelar
        que direcciones estan registradas.
      operationId: registrarse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudRegistro"
        required: true
      responses:
        "202":
          description: Solicitud recibida; revisa tu correo
        "400":
          description: Correo mal formado o clave que no cumple la politica
  /api/cuentas/invitacion:
    post:
      tags:
      - Cuentas
      summary: Aceptar una invitacion fijando la clave
      description: Deja la cuenta ACTIVA con el rol que eligio quien invito. No hace
        falta otra aprobacion.
      operationId: aceptar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudFijarClave"
        required: true
      responses:
        "204":
          description: Cuenta activa; ya puedes iniciar sesion
        "400":
          description: Enlace invalido o clave que no cumple la politica
  /api/cuentas/clave:
    post:
      tags:
      - Cuentas
      summary: Fijar la clave nueva con el token del enlace
      description: Cambia la clave y revoca todas las sesiones abiertas de esa cuenta.
      operationId: fijarClave
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudFijarClave"
        required: true
      responses:
        "204":
          description: Clave cambiada
        "400":
          description: "Enlace invalido, vencido o ya usado"
...
  /api/veedor/ingesta/salud:
    get:
      tags:
      - Veedor - Ingesta
      summary: "Salud de cada colector: última ejecución exitosa, ítems y tasa de\
        \ error"
      description: |-
        Lista vacía mientras el pipeline no haya corrido un ciclo. La telemetría vive en
        memoria del proceso, así que un reinicio la reinicia.
      operationId: salud
      responses:
        "200":
          description: Estado generado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaludColectorRespuesta"
        "401":
          description: Falta el token del veedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaludColectorRespuesta"
...
  /api/veedor/cortes/{id}:
    get:
      tags:
      - Veedor - Cortes
      summary: Consultar un corte por su identificador
      operationId: consultar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Corte encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "404":
          description: No existe un corte con ese id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/suscripciones/confirmar:
    get:
      tags:
      - Suscripciones
      summary: Confirmar la suscripción (doble opt-in)
      description: |-
        Enlace del correo de confirmación. El token es de un solo enlace, no de un solo uso:
        confirmarla dos veces no falla (RF013). Responde JSON o una página HTML de cortesía
        según el `Accept` de quien pide — el mismo enlace del correo abre bien tanto en un
        navegador como desde un cliente de API.
      operationId: confirmar_2
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      - name: Accept
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Suscripción confirmada
          content:
            application/json:
              schema:
                type: object
            text/html:
              schema:
                type: object
        "400":
          description: "Token inválido, inexistente o de una suscripción ya cancelada"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    SesionVeedor:
      type: object
      properties:
        token:
          type: string
          description: "Token JWT. Se envia como 'Authorization: Bearer <token>'"
        usuarioId:
          type: string
        nombre:
          type: string
        correo:
          type: string
        rol:
          type: string
          description: "ADMIN, VEEDOR u OBSERVADOR"
        permisos:
          type: array
          description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
            \ revocados"
          items:
            type: string
            description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
              \ revocados"
        alcance:
          type: string
          description: |-
            COMPLETO, o ALTA_SEGUNDO_FACTOR cuando la cuenta es ADMIN y todavia no dio de alta su
            TOTP. Con ese alcance el token solo sirve para /api/veedor/segundo-factor.
      description: "Sesion emitida para el panel del veedor (RNF011: expira en 8 horas)"
    SolicitudCodigo:
      type: object
      properties:
        codigo:
          minLength: 1
          type: string
      description: Codigo de 6 digitos de la app de autenticacion
...
    SolicitudCorte:
      required:
      - finPrometido
      - inicio
      type: object
      properties:
        sectoresAfectados:
          minItems: 1
          type: array
          description: Identificadores de los sectores afectados
          example:
          - manga
          - bocagrande
          items:
            type: string
            description: Identificadores de los sectores afectados
            example: "[\"manga\",\"bocagrande\"]"
        inicio:
          type: string
          format: date-time
        finPrometido:
          type: string
          format: date-time
        causa:
          minLength: 1
          type: string
          example: Mantenimiento planta El Bosque
      description: Registro de un corte oficial por el veedor (RF016)
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    CorteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectoresAfectados:
          type: array
          items:
            type: string
        inicio:
          type: string
          format: date-time
        finPrometido:
          type: string
          format: date-time
        finReal:
          type: string
          description: Nulo mientras el corte sigue abierto
          format: date-time
        causa:
          type: string
        origen:
          type: string
        estado:
          type: string
      description: Corte oficial (RF016-RF017)
...
    CoordenadaDTO:
      required:
      - latitud
      - longitud
      type: object
      properties:
        latitud:
          maximum: 90
          exclusiveMaximum: false
          minimum: -90
          exclusiveMinimum: false
          type: number
          format: double
        longitud:
          maximum: 180
          exclusiveMaximum: false
          minimum: -180
          exclusiveMinimum: false
          type: number
          format: double
      description: "Coordenada GPS del reporte, solo cuando el usuario la autoriza\
        \ (RF007)"
    SolicitudReporte:
      type: object
      properties:
        sectorId:
          minLength: 1
          type: string
          description: Identificador del sector reportado
          example: bocagrande
        tipo:
          minLength: 1
          type: string
          description: "SIN_AGUA, PRESION_BAJA o SERVICIO_RESTABLECIDO"
          example: SIN_AGUA
        huella:
          minLength: 1
          type: string
          description: |-
            Huella anónima del dispositivo (ADR-007) — no es una cuenta ni un identificador
            personal. El cliente la genera una vez (p. ej. un UUID persistido en el dispositivo,
            hasheado) y la reutiliza en cada reporte; es lo único que permite RF006 (límite de
            reportes por dispositivo) sin pedir registro.
        coordenada:
          $ref: "#/components/schemas/CoordenadaDTO"
      description: Reporte ciudadano sin registro (RF005-RF008)
    ReporteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectorId:
          type: string
        tipo:
          type: string
        timestamp:
          type: string
          format: date-time
        fotoUrl:
          type: string
        confirmaciones:
          type: integer
          format: int32
      description: Reporte ciudadano registrado
    SolicitudConfirmar:
      type: object
      properties:
        huella:
          maxLength: 128
          minLength: 32
          type: string
          description: Huella hash del dispositivo del usuario que confirma (ADR-007)
          example: 3a7b9c...
      description: Solicitud para confirmar un reporte ciudadano por otro vecino
...
    SolicitudRestablecer:
      type: object
      properties:
        correo:
          minLength: 1
          type: string
      description: "Pedir el enlace de restablecimiento. Responde siempre 202, exista\
        \ o no la cuenta."
    SolicitudRegistro:
      type: object
      properties:
        correo:
          minLength: 1
          type: string
        nombre:
          maxLength: 80
          minLength: 2
          type: string
          description: Nombre con el que apareceras en la auditoria del panel
        clave:
          minLength: 1
          type: string
          description: Minimo 12 caracteres. La politica completa vive en ClaveEnClaro.
      description: "Solicitud de acceso al panel. No concede nada: exige verificar\
        \ el correo y que un ADMIN apruebe."
    SolicitudFijarClave:
      type: object
      properties:
        token:
          minLength: 1
          type: string
          description: Token que venia en el enlace del correo
        clave:
          minLength: 1
          type: string
      description: Fijar clave desde un enlace de un solo uso (invitacion o restablecimiento)
...
    SaludColectorRespuesta:
      type: object
      properties:
        nombre:
          type: string
          example: acuacar
        ultimaEjecucionExitosa:
          type: string
          description: Nulo si el colector todavía no ha completado un ciclo con éxito
          format: date-time
          nullable: true
        ultimoFallo:
          type: string
          description: Nulo si nunca ha fallado
          format: date-time
          nullable: true
        motivoDelUltimoFallo:
          type: string
          description: "Mensaje del último fallo, para diagnosticar sin entrar al\
            \ servidor"
          nullable: true
        itemsProcesados:
          type: integer
          description: Documentos traídos desde que arrancó el proceso
          format: int64
        tasaDeError:
          type: number
          description: "Entre 0 y 1, sobre los ciclos corridos desde que arrancó el\
            \ proceso"
          format: double
        fallosConsecutivos:
          type: integer
          description: "Ciclos seguidos fallando. Desde 3, el colector se reporta\
            \ caído en /actuator/health"
          format: int32
      description: Estado de salud de un colector de la ingesta automatizada
...